productportal added

+++ ProductApp/Models/Customers.cs
@@ -12,7 +12,7 @@ namespace ProductApp.Models
         public string LastName { get; set; }
         public string Email { get; set; }
         public string Password { get; set; }
-        public int MobileNumber { get; set; }
+        public long MobileNumber { get; set; }
         public int Gender { get; set; }
         public string Address { get; set; }
         public DateTime DOB { get; set; }
